requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
Pillow>=10.0.0
selenium>=4.15.0
openai>=1.3.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
